--- a/tests/_adapt.py
+++ b/tests/_adapt.py
@@ -7,6 +7,7 @@ IS_GAE = "datastore" in DEFAULT_URI
 IS_MONGODB = "mongodb" in DEFAULT_URI
 IS_POSTGRESQL = 'postgres' in DEFAULT_URI
 IS_SQLITE = 'sqlite' in DEFAULT_URI
+IS_MSSQL = 'mssql' in DEFAULT_URI
 
 def drop(table, cascade=None):
     if NOSQL and not (IS_MONGODB):