According to my results, this PR slightly increases BasicStorage performance without breaking current unitests
```
import timeit
from pydal.helpers.classes import BasicStorage
bs = BasicStorage()
def bs_set():
    bs['a'] = 'pydal'  #or bs.a = 'pydal'

print timeit.timeit(bs_set, number=10000000)
```
 |Master pydal (s)| This PR (s)
----------|----------|-----------
setitem| 4.73 | 2.55
setattr | 4.73 |  1.37